% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{approx_popn}
\alias{approx_popn}
\title{Example approximate population (modelled after ACS)}
\format{
A data frame with columns:
\describe{
\item{age2}{Age group}
\item{gender}{Gender}
\item{pet_pref}{Pet preference}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in CreateTestData.R
}
\usage{
approx_popn
}
\description{
Example approximate population (modelled after ACS)
}
\examples{

 approx_popn

}
\keyword{datasets}
