% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{approx_voters_popn}
\alias{approx_voters_popn}
\title{Simulated approximate population data for the Shape World survey}
\format{
A data frame with columns:

\describe{
\item{age_group}{Age group}
\item{gender}{Gender}
\item{vote_for}{Party voted for in the 2016 election}
\item{education}{Highest grade of education completed}
\item{state}{State}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in ShapeWorldVoters.R
}
\usage{
approx_voters_popn
}
\description{
Simulated approximate population data for the \link[=shape_survey]{Shape World}
voters survey (modeled after ACS).
}
\examples{
approx_voters_popn

}
\keyword{datasets}
