% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_wtd_ests}
\alias{create_wtd_ests}
\title{Create weighted estimates using the survey package}
\usage{
create_wtd_ests(fit_obj, outcome, by = NULL)
}
\arguments{
\item{outcome}{The variable we are estimating}

\item{by}{The grouping variables}

\item{fit}{An \link[R6:R6Class]{R6} SurveyFit object}
}
\value{
A table of size nlevels by 3 (level, estimate, sd).
}
\description{
Create weighted estimates using the survey package
}
