% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feline_survey}
\alias{feline_survey}
\title{Example survey for package testing}
\format{
A data frame with columns:
\describe{
\item{age1}{Age group}
\item{gender}{Gender}
\item{pet_own}{Pet ownership}
\item{y}{Example binary outcome variable}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in CreateTestData.R
}
\usage{
feline_survey
}
\description{
Example survey for package testing
}
\examples{
str(feline_survey)

}
\keyword{datasets}
