% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feline_survey_haven}
\alias{feline_survey_haven}
\title{feline_survey with some variables stored in haven-labelled class.}
\format{
A data frame with columns:
\describe{
\item{age1}{Age group; haven-labelled}
\item{gender}{Gender; haven-labbeled}
\item{pet_own}{Pet ownership; haven-labelled}
\item{y}{Example binary outcome variable}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in CreateTestData.R
}
\usage{
feline_survey_haven
}
\description{
feline_survey with some variables stored in haven-labelled class.
}
\examples{
str(feline_survey_haven)

}
\keyword{datasets}
