% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bring_together_survey_and_poststrat.R
\name{go_get_em}
\alias{go_get_em}
\title{bring_together_survey_and_poststrat}
\usage{
go_get_em(some_survey_data, survey_variable_of_interest, aligned_levels_type)
}
\arguments{
\item{survey_variable_of_interest}{A csv specifying how the levels in each align.}

\item{survey_data}{A tibble of individuals and responses.}

\item{aligned_levels}{A csv specifying how the levels in each align.}
}
\value{
aligned_survey A modified version of the survey dataset that has its levels aligned with the poststratification dataset.
}
\description{
This takes a survey dataset and a post-stratification dataset and brings them together.
}
\details{
Bring together the survey and post-stratification datasets
}
\examples{
TBD

}
