% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_survey_popn.R
\name{map_survey_popn}
\alias{map_survey_popn}
\title{map_survey_popn}
\usage{
map_survey_popn(survey_answers, popn_answers, survey_name, popn_name,
  survey_map = NULL)
}
\arguments{
\item{popn_name}{}
}
\value{
A mapping object
}
\description{
Helps map responses in surveys
}
\details{
Map survey answers to population answers
}
\examples{
survey_answers <- c('18-25','26-35','36-45','46-55','56-65','66-75','76-90')
popn_answers <- c('18-35','18-35','36-55','36-55','56-65','66+','66+')
test_map <- map_survey_popn(survey_answers, popn_answers, "q1","q1alt")

survey_answers2 <- c('M','F')
popn_answers2 <- c('Male','Female')
test_map <- map_survey_popn(survey_answers2, popn_answers2, "q2","q2alt", survey_map = test_map)

}
