% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpkit-package.R
\docType{package}
\name{mrpkit-package}
\alias{mrpkit}
\alias{mrpkit-package}
\title{mrpkit: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
