% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nlp_survey}
\alias{nlp_survey}
\title{A Simulated Data of NLP Voters Survey}
\format{
A data frame with columns:

\describe{
\item{age}{Age group}
\item{gender}{Gender}
\item{vote_for}{The party that the respondent vote for in the 2016 election}
\item{highest_educ}{Highest grade of education that is completed}
\item{state}{State}
\item{y}{Preference to NLP in 2020 election}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in NeverlandVoters.R
}
\usage{
nlp_survey
}
\description{
This is a generated data shows the survey result of the preference
to Neverland Labor Party (NLP)
along with several demographic variables to demonstrate the usage of the package.
}
\examples{

 nlp_survey

}
\keyword{datasets}
