% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_survey.R
\name{print_survey}
\alias{print_survey}
\title{print_survey}
\usage{
print_survey(x, ..., verbose = TRUE)
}
\arguments{
\item{SurveyObj}{A survey object.}
}
\value{
A neat summmary.
}
\description{
Prints the survey object in a neat format
}
\details{
Print survey object summary
}
\examples{
test_obj <- SurveyObj(survey_data = data.frame(a=c(1,2),b=c(3,4)),
questions = c("q1","q2"),
answers = list(c('a','b'),c("c",'d')))
PrintSurvey(test_obj)

}
