% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_prepare_survey.R
\name{read_and_prepare_survey}
\alias{read_and_prepare_survey}
\title{read_and_prepare_survey}
\usage{
read_and_prepare_survey(dataset_of_interest, desired_variables,
  correct_class)
}
\arguments{
\item{dataset}{A CSV file in which every row is a response and every column is a variable.}

\item{variables}{A list that specifies the necessary columns.}

\item{class}{A list that specifies the class of each of the necessary columns.}
}
\value{
survey_dataset A dataset that is reduced to essentials and properly formatted.
}
\description{
This gets a CSV and checks whether the list of the variables is correct.
}
\details{
Read and prepare survey data
}
\examples{
TBD

}
