% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape_survey}
\alias{shape_survey}
\title{Simulated Survey Data for Shape World}
\format{
A data frame with columns:

\describe{
\item{age}{Age group}
\item{gender}{Gender}
\item{vote_for}{Party voted for in the 2016 election}
\item{highest_educ}{Highest grade of education completed}
\item{state}{State}
\item{y}{Preference for BP over CP in 2020 election}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in ShapeWorldVoters.R
}
\usage{
shape_survey
}
\description{
This is a simulated survey of whether Shape World voters prefer the Box Party
(BP) or the Circle Party (CP) in the 2020 election. The data also include
several demographic variables and the party the respondent voted for in the
previous election. We use this data throughout the documentation to
demonstrate the usage of the package.
}
\examples{

 shape_survey

}
\keyword{datasets}
