% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape_survey_haven}
\alias{shape_survey_haven}
\title{Simulated Survey Data for Shape World - Haven-labelled Version}
\format{
A data frame with columns:

\describe{
\item{age}{Age group; haven-labelled}
\item{gender}{Gender; haven-labelled}
\item{vote_for}{Party voted for in the 2016 election; haven-labelled}
\item{highest_educ}{Highest grade of education completed; haven-labelled}
\item{state}{State; haven-labelled}
\item{y}{Preference for BP over CP in 2020 election}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in ShapeWorldVoters.R
}
\usage{
shape_survey_haven
}
\description{
This is the shape_survey data with some variables saved in haven-labelled class.
}
\examples{

 shape_survey_haven

}
\keyword{datasets}
