% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_posterior_epred.R
\name{sim_posterior_epred}
\alias{sim_posterior_epred}
\title{sim_posterior_epred}
\usage{
sim_posterior_epred(object, newdata, nsamples = 4000)
}
\arguments{
\item{object}{Model of type glmer class and binomial family}

\item{newdata}{Data to predict for}

\item{nsamples}{Number of samples}
}
\description{
Helper function to create predictive samples from a glmer class model
Takes in model class and newdata to predict for
Return a matrix with rows that represent posterior samples and columns that represent
data that is being predicted for
Only for binomial family models.
}
