% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunction.R
\name{subtract}
\alias{subtract}
\title{Subtract together two numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Subtract together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
