% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyData.R
\name{SurveyData}
\alias{SurveyData}
\title{SurveyData objects}
\description{
An \link[R6:R6Class]{R6} \code{SurveyData} object represents a survey and its metadata.
The survey itself is a data frame.
The survey metatdata consists of the following:
\itemize{
\item per-column questions: a list of strings
\item per-column allowed response values: a list of character vectors
\item per-column survey weights: a vector of numeric weights
\item survey design: a string that specifies the survey design using \pkg{lme4}
style formula syntax.
}

Examples of survey designs include:
\itemize{
\item \code{~.}: a random sample
\item \code{~ (1|cluster)}: a one stage cluster sample
\item \code{~ stratum}: a stratified sample
}
}
\examples{
feline_prefs <- SurveyData$new(
  feline_survey[,c("age1","gender","pet_own","y")],
  questions = c(
    "Please identify your age group",
    "Please select your gender",
    "Which pet do you own?",
    "Response"
  ),
  responses = list(
    levels(feline_survey$age1),
    levels(feline_survey$gender),
    levels(feline_survey$pet_own),
    c("no","yes")
  ),
  weights = feline_survey$wt,
  design = formula("~.")
)
feline_prefs$print()

popn_obj <- SurveyData$new(
  approx_popn[,c("age2","gender","pet_pref")],
  questions = c(
    "Which age group are you?",
    "Gender?",
    "Which pet would you like to own?"
  ),
  responses = list(
    levels(approx_popn$age2),
    levels(approx_popn$gender),
    levels(approx_popn$pet_pref)
  ),
  weights = approx_popn$wt,
  design = formula("~.")
)
popn_obj$print()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyData$new()}}
\item \href{#method-n_obs}{\code{SurveyData$n_obs()}}
\item \href{#method-n_questions}{\code{SurveyData$n_questions()}}
\item \href{#method-print}{\code{SurveyData$print()}}
\item \href{#method-add_survey_data_column}{\code{SurveyData$add_survey_data_column()}}
\item \href{#method-add_mapped_data_column}{\code{SurveyData$add_mapped_data_column()}}
\item \href{#method-generate_poststrat_data}{\code{SurveyData$generate_poststrat_data()}}
\item \href{#method-survey_data}{\code{SurveyData$survey_data()}}
\item \href{#method-poststrat}{\code{SurveyData$poststrat()}}
\item \href{#method-mapped_data}{\code{SurveyData$mapped_data()}}
\item \href{#method-questions}{\code{SurveyData$questions()}}
\item \href{#method-responses}{\code{SurveyData$responses()}}
\item \href{#method-weights}{\code{SurveyData$weights()}}
\item \href{#method-design}{\code{SurveyData$design()}}
\item \href{#method-clone}{\code{SurveyData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$new(
  survey_data,
  questions = character(0),
  responses = list(),
  weights = numeric(),
  design = as.formula("~.")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_obs"></a>}}
\if{latex}{\out{\hypertarget{method-n_obs}{}}}
\subsection{Method \code{n_obs()}}{
Number of observations in the survey data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$n_obs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_questions"></a>}}
\if{latex}{\out{\hypertarget{method-n_questions}{}}}
\subsection{Method \code{n_questions()}}{
Number of survey questions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$n_questions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the survey data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_survey_data_column"></a>}}
\if{latex}{\out{\hypertarget{method-add_survey_data_column}{}}}
\subsection{Method \code{add_survey_data_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$add_survey_data_column(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_mapped_data_column"></a>}}
\if{latex}{\out{\hypertarget{method-add_mapped_data_column}{}}}
\subsection{Method \code{add_mapped_data_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$add_mapped_data_column(name, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_poststrat_data"></a>}}
\if{latex}{\out{\hypertarget{method-generate_poststrat_data}{}}}
\subsection{Method \code{generate_poststrat_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$generate_poststrat_data(grouping_vars)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survey_data"></a>}}
\if{latex}{\out{\hypertarget{method-survey_data}{}}}
\subsection{Method \code{survey_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$survey_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poststrat"></a>}}
\if{latex}{\out{\hypertarget{method-poststrat}{}}}
\subsection{Method \code{poststrat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$poststrat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_data}{}}}
\subsection{Method \code{mapped_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$mapped_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-questions"></a>}}
\if{latex}{\out{\hypertarget{method-questions}{}}}
\subsection{Method \code{questions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$questions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-responses"></a>}}
\if{latex}{\out{\hypertarget{method-responses}{}}}
\subsection{Method \code{responses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$responses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-weights"></a>}}
\if{latex}{\out{\hypertarget{method-weights}{}}}
\subsection{Method \code{weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$weights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-design"></a>}}
\if{latex}{\out{\hypertarget{method-design}{}}}
\subsection{Method \code{design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
