% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyFit.R
\name{SurveyFit}
\alias{SurveyFit}
\title{SurveyFit}
\description{
An \link[R6:R6Class]{R6} \code{SurveyFit} object stores a fitted model
object and provides methods for generating predicted probabilities for all
poststrat cells, generating population and group estimates, and visualizing
results.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyFit$new()}}
\item \href{#method-fit}{\code{SurveyFit$fit()}}
\item \href{#method-map}{\code{SurveyFit$map()}}
\item \href{#method-formula}{\code{SurveyFit$formula()}}
\item \href{#method-print}{\code{SurveyFit$print()}}
\item \href{#method-population_predict}{\code{SurveyFit$population_predict()}}
\item \href{#method-aggregate}{\code{SurveyFit$aggregate()}}
\item \href{#method-summary}{\code{SurveyFit$summary()}}
\item \href{#method-plot}{\code{SurveyFit$plot()}}
\item \href{#method-clone}{\code{SurveyFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SurveyFit} object. This method is called
internally by the \verb{$fit()} method of the \code{\link{SurveyMap}} object and does
not need to be called directly by the user.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$new(fit, map, formula)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit}}{A fitted model object.}

\item{\code{map}}{A \code{\link{SurveyMap}} object.}

\item{\code{formula}}{A formula object representing the model ran.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Access the fitted model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\if{latex}{\out{\hypertarget{method-map}{}}}
\subsection{Method \code{map()}}{
Access the SurveyMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formula"></a>}}
\if{latex}{\out{\hypertarget{method-formula}{}}}
\subsection{Method \code{formula()}}{
Access the SurveyMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Call the fitted model object's print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments to pass the print method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-population_predict"></a>}}
\if{latex}{\out{\hypertarget{method-population_predict}{}}}
\subsection{Method \code{population_predict()}}{
Use fitted model to add predicted probabilities to post-stratification dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$population_predict(..., fun = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments other than the fitted model object and
poststratification data frame to pass to \code{fun}.}

\item{\code{fun}}{The function to use to generate the predicted probabilities.
This should only be specified if you used a model fitting function
not natively supported by \pkg{mrpkit}.
For models fit using \pkg{rstanarm}, \pkg{brms}, or \pkg{lme4}, \code{fun}
is handled automatically. If \code{fun} is specified then:
\itemize{
\item the first argument should be the fitted model object
\item the second argument should be the poststratification data frame
\item it can take an arbitrary number of other arguments
\item the returned object should match the specifications in the 'Returns'
section below in order to be compatible with subsequent methods
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with rows corresponding to poststratification cells and
columns corresponding to posterior samples (or approximate ones
in the case of \pkg{lme4} models).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
Aggregate estimates to the population level or by level of a grouping variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$aggregate(poststrat_estimates, by = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{poststrat_estimates}}{The object returned by \code{population_predict}.}

\item{\code{by}}{Optionally a string specifying a grouping variable. If
specified the aggregation will happen by level of the named variable.
If not specified population-level estimates will be computed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame. If \code{by} is not specified then the data frame will
have number of rows equal to the number of posterior draws. If \code{by} is
specified the data frame will have number of rows equal to the number
of posterior draws times the number of levels of the \code{by} variable,
and there will be an extra column indicating which level of the \code{by}
variable each row corresponds to.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Creates a set of summary statistics of the mrp estimate,
and corresponding weighted and raw data estimates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$summary(aggregated_estimates)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregated_estimates}}{The data frame returned by the \verb{aggregate() method}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame that consists of a minimum three rows with the raw, MRP
and weighted estimates, plus an estimate of standard error. If the aggregated estimates
were specified with a \code{by} argument (indicating sub population or small area estimates),
then produces a dataframe with number of rows equal to three times the number of small areas.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot takes the aggregated MRP estimates and produces a quick
visualization of total and sub-population estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$plot(aggregated_estimates, additional_stats = c("wtd", "raw"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggregated_estimates}}{The data frame returned by \code{aggregate}.}

\item{\code{additional_stats}}{A vector that specifies which of three additional
stats ("wtd", "raw", "mrp", "none") should be included on the plot. The
default value is to include the weighted estimate (wtd) and raw data
mean (raw), but an analogous bar for MRP (mrp) can be added using the
posterior mean and sd. The sd for the weighted estimate uses the survey
design and the \pkg{survey} package, whilst the raw estimate is a
direct mean and binomial sd of the binary responses. Uncertainty is included on
violin plots but not on density plots. Intervals are 95\% CI.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object that is either a violin plot (if aggregated estimates is
at a small area level) or density plot (if aggregated estimates is at the population
level).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
