% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap_alt.R
\docType{data}
\name{SurveyMap_alt}
\alias{SurveyMap_alt}
\title{SurveyMap}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SurveyMap_alt
}
\description{
A `SurveyMap_alt` object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.   The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\examples{
q1 <- question$new(name = "age", col_names = c("age1","age_group"),
values_map = list("18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
        "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"))
q2 <- question$new(name = "pet", col_names = c("pet_pref","pet_own"),
values = list("cat" = "cat", "kitten" = "cat","dog" = "dog","puppy" = "dog"))
q3 <- question$new(name = "gender", col_names = c("gender","gender"),
values = data.frame("male" = "m","female" = "f", "nonbinary" = "nb"))
tmp_map <- SurveyMap_alt$new(q1)
print(tmp_map)
tmp_map$add(q3,q2)
print(tmp_map)
tmp_map$delete(q3)
print(tmp_map)
tmp_map$delete("pet")
print(tmp_map)
tmp_map$add(q2)
print(tmp_map)
tmp_map$replace(q1,q3)
print(tmp_map)
}
\keyword{datasets}
