% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyObj.R
\name{SurveyObj}
\alias{SurveyObj}
\title{An S4 class to represent a survey}
\usage{
SurveyObj(survey_data, questions = NA, answers = NA)
}
\description{
An S4 class to represent a survey
}
\section{Slots}{

\describe{
\item{\code{survey_data}}{A dataframe}

\item{\code{questions}}{A vector of questions}
}}

