% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyObj.R
\docType{data}
\name{SurveyObject}
\alias{SurveyObject}
\alias{SurveyObj}
\title{SurveyObject}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SurveyObj
}
\description{
A `SurveyObject` represents a survey and its metadata.
The survey itself is a dataframe.
The survey metatdata consists of
 - per-column questions - a list of strings
 - per-column allowed response values - a list of lists
 - per-column survey weights, - a vector of numeric weights
 - a string that specifies the survey design using lmer forumula syntax

Examples of survey design include:
~. a random sample, ~ (1|cluster), a one stage cluster sample, ~ stratum, a stratified sample
}
\examples{
feline_prefs = SurveyObj$new(feline_survey)
}
\keyword{datasets}
