% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyQuestion.R
\docType{data}
\name{SurveyQuestion}
\alias{SurveyQuestion}
\alias{question}
\title{SurveyQuestion}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
question
}
\description{
A \code{SurveyQuestion} object holds the mapping for one question
or demographic between the survey and population dataset.
The name is the name of the underlying construct.
The col_names maps the survey column name to the population column name.
The values map the survey response values to the population values.
If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
The ordering should be sample first and then population
}
\examples{

q1 <- question$new(
  name = "age",
  col_names = c("age1","age2"),
  values_map = list(
    "18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
    "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"
  )
)
q2 <- question$new(
  name = "pet",
  col_names = c("pet_pref","pet_own"),
  values_map = list("cat" = "cat", "kitten" = "cat","dog" = "dog","dog" = "puppy")
)
q3 <- question$new(
  name = "gender",
  col_names = c("gender","gender"),
  values_map = data.frame("male" = "m","female" = "f", "nonbinary" = "nb")
)

}
\keyword{datasets}
