% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{approx_voters_popn}
\alias{approx_voters_popn}
\title{A Simulated Data of NLP Voters Approximate Population (Modelled after ACS)}
\format{
A data frame with columns:

\describe{
\item{age_group}{Age group}
\item{gender}{Gender}
\item{vote_for}{The party that being voted in the 2016 election}
\item{education}{Highest grade of education that is completed}
\item{state}{State}
\item{wt}{Raked and trimmed survey weight}
}
}
\source{
Created in NeverlandVoters.R
}
\usage{
approx_voters_popn
}
\description{
This is a generated data shows the approximate population of the
Neverland Labor Party (NLP) Survey to demonstrate the usage of the package.
}
\examples{

 approx_voters_popn

}
\keyword{datasets}
