% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrpkit-package.R
\docType{package}
\name{mrpkit-package}
\alias{mrpkit}
\alias{mrpkit-package}
\title{mrpkit: Multilevel Regression with Post-Stratification}
\description{
Conducts multilevel regression with post-stratification in an opinionated and specific way. Users specify two datasets (sample and population), establish a correspondence between them, and then conduct analysis.
}
\author{
\strong{Maintainer}: Lauren Kennedy \email{lauren.kennedy1@monash.edu}

Authors:
\itemize{
  \item Jonah Gabry \email{jsg2201@columbia.edu}
  \item Dewi Amaliah
  \item Rohan Alexander \email{rohan.alexander@utoronto.ca}
}

Other contributors:
\itemize{
  \item Mitzi Morris [contributor]
}

}
\keyword{internal}
