% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_survey.R
\name{print_survey}
\alias{print_survey}
\title{print_survey}
\usage{
print_survey(x, ..., verbose = TRUE)
}
\arguments{
\item{SurveyObj}{or SurveyMap}
}
\value{
A neat summmary.
}
\description{
Prints the survey object or survey mapin a neat format
}
\details{
Print survey object summary
}
\examples{
test_obj <- SurveyObj(survey_data = data.frame(a=c(1,2),b=c(3,4)),
questions = c("q1","q2"),
answers = list(c('a','b'),c("c",'d')),
design =  "~.")
print_survey(test_obj)

test_obj_cluster <- SurveyObj(survey_data = data.frame(a=c(1,2),b=c(3,4)),
questions = c("q1","q2"),
answers = list(c('a','b'),c("c",'d')),
design =  as.formula("~ (1|school)"))
print_survey(test_obj_cluster)

test_obj_stratified <- SurveyObj(survey_data = data.frame(a=c(1,2),b=c(3,4)),
questions = c("q1","q2"),
answers = list(c('a','b'),c("c",'d')),
design =  as.formula("~ geo_strata"))
print_survey(test_obj_stratified)

survey_answers <- c('18-25','26-35','36-45','46-55','56-65','66-75','76-90')
popn_answers <- c('18-35','18-35','36-55','36-55','56-65','66+','66+')
test_map <- map_survey_popn(survey_answers, popn_answers, "q1","q1alt")
survey_answers2 <- c('M','F')
popn_answers2 <- c('Male','Female')
test_map <- map_survey_popn(survey_answers2, popn_answers2, "q2","q2alt", survey_map = test_map)
print_survey(test_map)

}
