% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_posterior_epred.R
\name{sim_posterior_epred}
\alias{sim_posterior_epred}
\title{Generate approximate samples of predicted probabilities from a binomial
\code{glmerMod} object.}
\usage{
sim_posterior_epred(object, newdata, nsamples = 4000)
}
\arguments{
\item{object}{A binomial model fit using \code{lme4::glmer()}.}

\item{newdata}{The data to use.}

\item{nsamples}{Number of samples to generate.}
}
\value{
A matrix with \code{nrow(newdata)} rows and \code{nsamples} columns.
}
\description{
Generate approximate samples of predicted probabilities from a binomial
\code{glmerMod} object.
}
\examples{

}
