% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyData.R
\name{SurveyData}
\alias{SurveyData}
\title{SurveyData objects}
\description{
An \link[R6:R6Class]{R6} \code{SurveyData} object represents a survey and its
metadata. The survey itself is a data frame containing all data from the
survey. The \code{SurveyData} object also includes the survey questions and
responses (if left empty these will just be the column and factor level
names). To enable weighted comparisons, survey weights and a survey design
can be specified, with the survey design specified using \pkg{survey} package
notation.
}
\examples{

## ------------------------------------------------
## Method `SurveyData$new`
## ------------------------------------------------

# Example sample data
head(shape_survey)

# SurveyData object for sample data
box_prefs <- SurveyData$new(
  data = shape_survey,
  questions = list(
    age = "Please identify your age group",
    gender = "Please select your gender",
    vote_for = "Which party did you vote for in the 2018 election?",
    y = "If today is the election day, would you vote for the Box Party?"
  ),
  responses = list(
    age = levels(shape_survey$age),
    gender = levels(shape_survey$gender),
    # Here we use a data frame for the responses because the levels
    # in the data are abridged versions of the actual responses.
    # This can be useful when surveys have brief/non descriptive responses.
    vote_for = data.frame(
      data = levels(shape_survey$vote_for),
      asked = c("Box Party Faction A", "Box Party Faction B",
                "Circle Party Coalition", "Circle Party")
    ),
    y = c("no","yes")
  ),
  weights = "wt",
  design = list(ids =~1)
)
box_prefs$print()
box_prefs$n_questions()


# Example population data
head(approx_voters_popn)

# SurveyData object for population if estimated from large survey
popn_obj1 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  # order doesn't matter (gender before age here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  ),
  weights = "wt" # use the wt column from approx_voters_popn data
)

# SurveyData object for population if poststratification matrix already known
library(dplyr)
popn_ps <- approx_voters_popn \%>\%
  group_by(age_group,gender) \%>\%
  summarise(N_j = sum(wt))

popn_obj2 <- SurveyData$new(
  data = popn_ps,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(popn_ps$gender),
    age_group = levels(popn_ps$age_group)
  ),
  weights = "N_j"# use N_j column from popn_ps data
)

# SurveyData object for population if individual population data known:
# (pretend that approx_voters_popn is the full population)
popn_obj3 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  )
)
popn_obj1
popn_obj2
popn_obj3

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SurveyData-new}{\code{SurveyData$new()}}
\item \href{#method-SurveyData-n_obs}{\code{SurveyData$n_obs()}}
\item \href{#method-SurveyData-n_questions}{\code{SurveyData$n_questions()}}
\item \href{#method-SurveyData-print}{\code{SurveyData$print()}}
\item \href{#method-SurveyData-add_survey_data_column}{\code{SurveyData$add_survey_data_column()}}
\item \href{#method-SurveyData-survey_data}{\code{SurveyData$survey_data()}}
\item \href{#method-SurveyData-questions}{\code{SurveyData$questions()}}
\item \href{#method-SurveyData-responses}{\code{SurveyData$responses()}}
\item \href{#method-SurveyData-weights}{\code{SurveyData$weights()}}
\item \href{#method-SurveyData-design}{\code{SurveyData$design()}}
\item \href{#method-SurveyData-clone}{\code{SurveyData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-new"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SurveyData} object using an existing data
frame and other survey information. This method is used to create the
objects for both the sample and the population data. If a population is
approximated from a large survey (like the ACS or DHS), then the
package will enable the creation of a weighted poststratification
matrix. If the population is summarized as a poststratification matrix
already, then set the weights as the size in each cell $N_j$. If the
entire individual level population data is given, then weights should
be omitted and will be automatically set to 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$new(
  data,
  questions = list(),
  responses = list(),
  weights = numeric(),
  design = list(ids = ~1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing the survey data.}

\item{\code{questions, responses}}{Named lists containing the text of the survey
questions and the allowed responses, respectively. The names must
correspond to the names of variables in \code{data}. If these aren't
provided then they will be created internally using all factor,
character, and binary variables in \code{data}. Responses can also be
provided as a data frame with column names \code{"data"} and \code{"asked"},
which reflect the responses as coded in the data (\code{"data"}) and their
corresponding actual survey responses (\code{"asked"}). See \strong{Examples}.}

\item{\code{weights}}{Optionally, the name of a variable in \code{data} containing
survey weights.}

\item{\code{design}}{Optionally, a named list of arguments (except \code{weights} and
\code{data}) to pass to \code{survey::svydesign()} to specify the survey design.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{SurveyData} object that can be used in the creation of a
\code{\link{SurveyMap}} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example sample data
head(shape_survey)

# SurveyData object for sample data
box_prefs <- SurveyData$new(
  data = shape_survey,
  questions = list(
    age = "Please identify your age group",
    gender = "Please select your gender",
    vote_for = "Which party did you vote for in the 2018 election?",
    y = "If today is the election day, would you vote for the Box Party?"
  ),
  responses = list(
    age = levels(shape_survey$age),
    gender = levels(shape_survey$gender),
    # Here we use a data frame for the responses because the levels
    # in the data are abridged versions of the actual responses.
    # This can be useful when surveys have brief/non descriptive responses.
    vote_for = data.frame(
      data = levels(shape_survey$vote_for),
      asked = c("Box Party Faction A", "Box Party Faction B",
                "Circle Party Coalition", "Circle Party")
    ),
    y = c("no","yes")
  ),
  weights = "wt",
  design = list(ids =~1)
)
box_prefs$print()
box_prefs$n_questions()


# Example population data
head(approx_voters_popn)

# SurveyData object for population if estimated from large survey
popn_obj1 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  # order doesn't matter (gender before age here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  ),
  weights = "wt" # use the wt column from approx_voters_popn data
)

# SurveyData object for population if poststratification matrix already known
library(dplyr)
popn_ps <- approx_voters_popn \%>\%
  group_by(age_group,gender) \%>\%
  summarise(N_j = sum(wt))

popn_obj2 <- SurveyData$new(
  data = popn_ps,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(popn_ps$gender),
    age_group = levels(popn_ps$age_group)
  ),
  weights = "N_j"# use N_j column from popn_ps data
)

# SurveyData object for population if individual population data known:
# (pretend that approx_voters_popn is the full population)
popn_obj3 <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?"
  ),
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group)
  )
)
popn_obj1
popn_obj2
popn_obj3

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-n_obs"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-n_obs}{}}}
\subsection{Method \code{n_obs()}}{
Number of observations in the survey data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$n_obs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-n_questions"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-n_questions}{}}}
\subsection{Method \code{n_questions()}}{
Number of survey questions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$n_questions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-print"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the survey data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{SurveyData} object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-add_survey_data_column"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-add_survey_data_column}{}}}
\subsection{Method \code{add_survey_data_column()}}{
Add a column to the sample data. This is primarily
intended for internal use but may occasionally be useful.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$add_survey_data_column(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name, value}}{The name of the new variable (a string) and the
vector of values to add to the data frame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{SurveyData} object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-survey_data"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-survey_data}{}}}
\subsection{Method \code{survey_data()}}{
Access the data frame containing the sample data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$survey_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-questions"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-questions}{}}}
\subsection{Method \code{questions()}}{
Access the list of survey questions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$questions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-responses"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-responses}{}}}
\subsection{Method \code{responses()}}{
Access the list of allowed survey responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$responses()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-weights"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-weights}{}}}
\subsection{Method \code{weights()}}{
Access the survey weights
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$weights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-design"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-design}{}}}
\subsection{Method \code{design()}}{
Access the survey design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveyData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SurveyData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
