% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyData.R
\name{SurveyData}
\alias{SurveyData}
\title{SurveyData objects}
\description{
A \code{SurveyData} object represents a survey and its metadata. The survey itself
is a data frame. The survey metadata consists of the text of the survey
questions, the allowed response values, and optionally survey weights and a
survey design specification.
}
\examples{

head(feline_survey)
feline_prefs <- SurveyData$new(
  data = feline_survey,
  questions = list(
    age1 = "Please identify your age group",
    gender = "Please select your gender",
    pet_own = "Which pet do you own?",
    y = "Response"
  ),
  responses = list(
    age1 = levels(feline_survey$age1),
    gender = levels(feline_survey$gender),
    pet_own = levels(feline_survey$pet_own),
    y = c("no","yes")
  ),
  weights = feline_survey$wt,
  design = list(ids =~1)
)
feline_prefs$print()
feline_prefs$n_questions()

head(approx_popn)
popn_obj <- SurveyData$new(
  data = approx_popn,
  questions = list(
    age2 = "Which age group are you?",
    gender = "Gender?",
    pet_pref = "Which pet would you like to own?"
  ),
  # order doesn't matter (gender before age2 here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_popn$gender),
    age2 = levels(approx_popn$age2),
    pet_pref = levels(approx_popn$pet_pref)
  ),
  weights = approx_popn$wt
)
popn_obj$print()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyData$new()}}
\item \href{#method-n_obs}{\code{SurveyData$n_obs()}}
\item \href{#method-n_questions}{\code{SurveyData$n_questions()}}
\item \href{#method-print}{\code{SurveyData$print()}}
\item \href{#method-add_survey_data_column}{\code{SurveyData$add_survey_data_column()}}
\item \href{#method-add_mapped_data_column}{\code{SurveyData$add_mapped_data_column()}}
\item \href{#method-survey_data}{\code{SurveyData$survey_data()}}
\item \href{#method-mapped_data}{\code{SurveyData$mapped_data()}}
\item \href{#method-questions}{\code{SurveyData$questions()}}
\item \href{#method-responses}{\code{SurveyData$responses()}}
\item \href{#method-weights}{\code{SurveyData$weights()}}
\item \href{#method-design}{\code{SurveyData$design()}}
\item \href{#method-clone}{\code{SurveyData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SurveyData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$new(
  data,
  questions = list(),
  responses = list(),
  weights = numeric(),
  design = list(ids = ~1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame containing the survey data.}

\item{\code{questions, responses}}{Named lists containing the text of the
survey questions and the allowed responses, respectively. The names
must correspond to the names of variables in \code{data}. See
\strong{Examples}.}

\item{\code{weights}}{Optionally, a vector of survey weights.}

\item{\code{design}}{Optionally, a named list of arguments to pass to \code{survey::svydesign()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_obs"></a>}}
\if{latex}{\out{\hypertarget{method-n_obs}{}}}
\subsection{Method \code{n_obs()}}{
Number of observations in the survey data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$n_obs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_questions"></a>}}
\if{latex}{\out{\hypertarget{method-n_questions}{}}}
\subsection{Method \code{n_questions()}}{
Number of survey questions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$n_questions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the survey data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_survey_data_column"></a>}}
\if{latex}{\out{\hypertarget{method-add_survey_data_column}{}}}
\subsection{Method \code{add_survey_data_column()}}{
Add a column to the sample data. This is primarily
intended for internal use but may occasionally be useful.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$add_survey_data_column(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name, value}}{The name of the new variable (a string) and the
vector of values to add to the data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_mapped_data_column"></a>}}
\if{latex}{\out{\hypertarget{method-add_mapped_data_column}{}}}
\subsection{Method \code{add_mapped_data_column()}}{
Add a column to the mapped data. This is primarily
intended for internal use but may occasionally be useful.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$add_mapped_data_column(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name, value}}{The name of the new variable (a string) and the
vector of values to add to the data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survey_data"></a>}}
\if{latex}{\out{\hypertarget{method-survey_data}{}}}
\subsection{Method \code{survey_data()}}{
Access the data frame containing the sample data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$survey_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_data}{}}}
\subsection{Method \code{mapped_data()}}{
Access the data frame containing the mapped data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$mapped_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-questions"></a>}}
\if{latex}{\out{\hypertarget{method-questions}{}}}
\subsection{Method \code{questions()}}{
Access the list of survey questions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$questions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-responses"></a>}}
\if{latex}{\out{\hypertarget{method-responses}{}}}
\subsection{Method \code{responses()}}{
Access the list of allowed survey responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$responses()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-weights"></a>}}
\if{latex}{\out{\hypertarget{method-weights}{}}}
\subsection{Method \code{weights()}}{
Access the survey weights
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$weights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-design"></a>}}
\if{latex}{\out{\hypertarget{method-design}{}}}
\subsection{Method \code{design()}}{
Access the survey design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
