% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyFit.R
\name{SurveyFit}
\alias{SurveyFit}
\title{SurveyFit}
\description{
TO-DO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyFit$new()}}
\item \href{#method-fit}{\code{SurveyFit$fit()}}
\item \href{#method-map}{\code{SurveyFit$map()}}
\item \href{#method-print}{\code{SurveyFit$print()}}
\item \href{#method-predictify}{\code{SurveyFit$predictify()}}
\item \href{#method-collapsify}{\code{SurveyFit$collapsify()}}
\item \href{#method-visify}{\code{SurveyFit$visify()}}
\item \href{#method-clone}{\code{SurveyFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$new(fit, map)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Access the fitted model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The fitted model object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\if{latex}{\out{\hypertarget{method-map}{}}}
\subsection{Method \code{map()}}{
Access the SurveyMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$map()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \link{SurveyMap} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Call the fitted model object's print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optional arguments to pass the print method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predictify"></a>}}
\if{latex}{\out{\hypertarget{method-predictify}{}}}
\subsection{Method \code{predictify()}}{
Use fitted model to add predicted probabilities to post-stratification dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$predictify(fun = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{The function to use to generate the predicted probabilities.
This should only be specified if using a custom function, otherwise for
\pkg{rstanarm} and \pkg{brms} models \code{posterior_epred()} is
automatically used (with the result transposed) and for \pkg{lme4}
models the \code{\link[=sim_posterior_epred]{sim_posterior_epred()}} is used. If \code{fun} is a custom
function then the first argument should take in the fitted model object
and the second argument should take in the poststratification
(\code{newdata}) data frame. The function must return a matrix with rows
corresponding to the columns of the poststratification data and columns
corresponding to simulations.}

\item{\code{...}}{Arguments other than the fitted model and \code{newdata} data frame
to pass to \code{fun}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix with rows corresponding to poststrat cells and columns
corresponding to posterior samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collapsify"></a>}}
\if{latex}{\out{\hypertarget{method-collapsify}{}}}
\subsection{Method \code{collapsify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$collapsify(poststrat_fit, variable_aggr = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-visify"></a>}}
\if{latex}{\out{\hypertarget{method-visify}{}}}
\subsection{Method \code{visify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$visify(sae_preds)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
