% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.R
\docType{data}
\name{SurveyMap}
\alias{SurveyMap}
\title{SurveyMap}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SurveyMap
}
\description{
A `SurveyMap` object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.   The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\examples{
survey_values <- c('18-25','26-35','36-45','46-55','56-65','66-75','76-90')
popn_values <- c('18-35','18-35','36-55','36-55','56-65','66+','66+')
test_map <- SurveyMap$new("age_s", "age_p", survey_values, popn_values)

survey_values2 <- c('M','F')
popn_values2 <- c('Male','Female')
test_map$add_item("sex_s", "sex_p", survey_values2, popn_values2)

test2_map$new(c("age_s", "sex_s"), c("age_p", "sex_p"),
              c(list(survey_values, survey_values2), list(pop_values, pop_values2))
}
\keyword{datasets}
