% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.R
\name{SurveyMap}
\alias{SurveyMap}
\title{An S4 class to map survey classes together}
\usage{
SurveyMap(SurveyMap, survey_answer, popn_answer, survey_name, popn_name)
}
\description{
An S4 class to map survey classes together
}
\section{Slots}{

\describe{
\item{\code{survey_vars}}{A vector of colnames}

\item{\code{popn_vars}}{A vector of colnames}
}}

