% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.R
\name{SurveyMap}
\alias{SurveyMap}
\title{SurveyMap}
\description{
An \link[R6:R6Class]{R6} \code{SurveyMap} object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.  The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyMap$new()}}
\item \href{#method-print}{\code{SurveyMap$print()}}
\item \href{#method-add}{\code{SurveyMap$add()}}
\item \href{#method-delete}{\code{SurveyMap$delete()}}
\item \href{#method-replace}{\code{SurveyMap$replace()}}
\item \href{#method-validate}{\code{SurveyMap$validate()}}
\item \href{#method-mapping}{\code{SurveyMap$mapping()}}
\item \href{#method-tabulate}{\code{SurveyMap$tabulate()}}
\item \href{#method-fit}{\code{SurveyMap$fit()}}
\item \href{#method-item_map}{\code{SurveyMap$item_map()}}
\item \href{#method-sample}{\code{SurveyMap$sample()}}
\item \href{#method-population}{\code{SurveyMap$population()}}
\item \href{#method-poststrat_data}{\code{SurveyMap$poststrat_data()}}
\item \href{#method-mapped_sample_data}{\code{SurveyMap$mapped_sample_data()}}
\item \href{#method-mapped_population_data}{\code{SurveyMap$mapped_population_data()}}
\item \href{#method-clone}{\code{SurveyMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SurveyMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$new(sample, population, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{The \link{SurveyData} object corresponding to the sample data.}

\item{\code{population}}{The \link{SurveyData} object corresponding to the population data.}

\item{\code{...}}{\link{QuestionMap} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the mapping.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add new \link{QuestionMap}s.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The \link{QuestionMap}s to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete \link{QuestionMap}s.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$delete(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The \link{QuestionMap}s to delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace"></a>}}
\if{latex}{\out{\hypertarget{method-replace}{}}}
\subsection{Method \code{replace()}}{
Replace one \link{QuestionMap} with another.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$replace(old_question, new_question)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_question}}{The \link{QuestionMap} object to replace.}

\item{\code{new_question}}{The \link{QuestionMap} object to use instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the mapping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapping"></a>}}
\if{latex}{\out{\hypertarget{method-mapping}{}}}
\subsection{Method \code{mapping()}}{
The mapping method uses the given maps between questions to create new sample and population data
that has unified variable names (i.e., if the underlying construct is called \code{age}, both sample and population will have
an \code{age} column, even if in the the raw data both had different variable names). The method also unifies the levels of each
variable in the sample and population so that the maximum set of consistent levels is created. Names of these new levels are
made according the the sample level names. If multiple levels are combined, the new name will be existing levels seperated by a
\code{+}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate}{}}}
\subsection{Method \code{tabulate()}}{
Prepare the poststratification table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$tabulate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The variables to include.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$fit(fun, formula, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{The model fitting function to use. For example,
\code{fun=rstanarm::stan_glmer}, \code{fun=brms::brm}, \code{fun=lme4::glmer}.
If using a custom \code{fun} it must have a \code{data} argument that accepts
a data frame (like standard R modeling functions).}

\item{\code{formula}}{The model formula. Can be either a string or a formula
object.}

\item{\code{...}}{Arguments other than \code{formula} and \code{data} to pass to \code{fun}.
The data argument will be automatically specified internally.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{SurveyFit} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-item_map"></a>}}
\if{latex}{\out{\hypertarget{method-item_map}{}}}
\subsection{Method \code{item_map()}}{
Access the \code{item_map}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$item_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Access the \link{SurveyData} object containing the sample data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-population"></a>}}
\if{latex}{\out{\hypertarget{method-population}{}}}
\subsection{Method \code{population()}}{
Access the \link{SurveyData} object containing the population data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$population()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poststrat_data"></a>}}
\if{latex}{\out{\hypertarget{method-poststrat_data}{}}}
\subsection{Method \code{poststrat_data()}}{
Access the poststratification data frame created by the \code{tabulate} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$poststrat_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_sample_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_sample_data}{}}}
\subsection{Method \code{mapped_sample_data()}}{
Access the data frame containing the mapped sample data
created by the \code{mapping} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapped_sample_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_population_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_population_data}{}}}
\subsection{Method \code{mapped_population_data()}}{
Access the data frame containing the mapped population data
created by the \code{mapping} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapped_population_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
