% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.R
\name{SurveyMap}
\alias{SurveyMap}
\title{SurveyMap}
\description{
A \code{SurveyMap} object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.  The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\examples{
feline_prefs = SurveyObj$new(
  feline_survey[,c("age1","gender","pet_own","y")],
  questions = c("Please identify your age group","Please select your gender","Which pet do you own?", "Response"),
  responses = list(levels(feline_survey$age1),levels(feline_survey$gender),levels(feline_survey$pet_own),c("no","yes")),
  weights = feline_survey$wt,
  design = formula("~.")
)
popn_obj = SurveyObj$new(
  approx_popn[,c("age2","gender","pet_pref")],
  questions = c("Which age group are you?","Gender?","Which pet would you like to own?"),
  responses = list(levels(approx_popn$age2),levels(approx_popn$gender),levels(approx_popn$pet_pref)),
  weights = approx_popn$wt,
  design = formula("~.")
)
q1 <- question$new(
  name = "age",
  col_names = c("age1","age2"),
  values_map = list(
    "18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
    "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"
  )
)
q2 <- question$new(
  name = "pet",
  col_names = c("pet_own","pet_pref"),
  values_map = list("cat" = "cat", "kitten" = "cat","dog" = "dog","puppy" = "dog")
)
q3 <- question$new(
  name = "gender",
  col_names = c("gender","gender"),
  values_map = data.frame("male" = "m","female" = "f", "nonbinary" = "nb")
)

tmp_map <- SurveyMap$new(samp_obj = feline_prefs, popn_obj = popn_obj, q1)
print(tmp_map)
tmp_map$validate()
tmp_map$add(q3)
print(tmp_map)
tmp_map$delete(q3)
print(tmp_map)
tmp_map$add(q3)
tmp_map$delete("gender")
print(tmp_map)
tmp_map$add(q2)
print(tmp_map)
tmp_map$replace(q1,q3)
print(tmp_map)
tmp_map$add(q1)
print(tmp_map)
tmp_map$validate() # TODO: figure out why this errors
tmp_map$mapping()
tmp_map$tabulate("age") #Just use age in the poststrat matrix
tmp_map$tabulate() #Use all variables in the map


mod_fit_1 <- tmp_map$fit(
  fun = rstanarm::stan_glmer,
  formula = y ~ (1|age) + (1|gender),
  family = "binomial",
  refresh = 100,
  cores = 2
)

mod_fit_2 <- tmp_map$fit(
  fun = brms::brm,
  formula = y ~ (1|age) + (1|gender),
  family = "bernoulli",
  refresh = 100,
  cores = 2
)

poststrat_fit <- tmp_map$predictify(mod_fit_1) # predict in postrat matrix
-returns a matrix with rows as poststrat rows, columns as posterior samples, NEED TO DOCUMENT!!!
tmp_map$poststratify(poststrat_fit, variable = "age") # get an estimate for a particular level
plot1 <- tmp_map$visualize()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyMap$new()}}
\item \href{#method-print}{\code{SurveyMap$print()}}
\item \href{#method-add}{\code{SurveyMap$add()}}
\item \href{#method-delete}{\code{SurveyMap$delete()}}
\item \href{#method-replace}{\code{SurveyMap$replace()}}
\item \href{#method-validate}{\code{SurveyMap$validate()}}
\item \href{#method-mapping}{\code{SurveyMap$mapping()}}
\item \href{#method-tabulate}{\code{SurveyMap$tabulate()}}
\item \href{#method-fit}{\code{SurveyMap$fit()}}
\item \href{#method-predictify}{\code{SurveyMap$predictify()}}
\item \href{#method-clone}{\code{SurveyMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$new(samp_obj, popn_obj, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$add(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$delete(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace"></a>}}
\if{latex}{\out{\hypertarget{method-replace}{}}}
\subsection{Method \code{replace()}}{
Replace one survey question with another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$replace(old_question, new_question)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_question}}{The \link{SurveyQuestion} object to replace.}

\item{\code{new_question}}{The \link{SurveyQuestion} object to use instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapping"></a>}}
\if{latex}{\out{\hypertarget{method-mapping}{}}}
\subsection{Method \code{mapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate}{}}}
\subsection{Method \code{tabulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$tabulate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$fit(fun, formula, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{The model fitting function to use. For example,
\code{fun=rstanarm::stan_glmer} or \code{fun=brms::brm}.}

\item{\code{formula}}{The model formula. Can be either a string or a formula
object.}

\item{\code{...}}{Arguments other than \code{formula} and \code{data} to pass to the model
fitting function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predictify"></a>}}
\if{latex}{\out{\hypertarget{method-predictify}{}}}
\subsection{Method \code{predictify()}}{
Use fitted model to add predictions to post-stratification dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$predictify(fitted_model, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fitted_model}}{The name of the model that was fit. For example, \code{fit1}.}

\item{\code{...}}{Arguments other than \code{fitted_model} to pass to the prediction
function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
