% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.R
\name{SurveyMap}
\alias{SurveyMap}
\title{SurveyMap}
\description{
An \link[R6:R6Class]{R6} \code{SurveyMap} object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.  The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\examples{
library(dplyr)

head(shape_survey)
box_prefs <- SurveyData$new(
  data = shape_survey,
  questions = list(
    age = "Please identify your age group",
    gender = "Please select your gender",
    vote_for = "Which party did you vote for in the 2018 election?",
    y = "If today is the election day, would you vote for the Box Party?"
  ),
  responses = list(
    age = levels(shape_survey$age),
    gender = levels(shape_survey$gender),
    vote_for = levels(shape_survey$vote_for),
    y = c("no","yes")
  ),
  weights = "wt",
  design = list(ids =~1)
)
box_prefs$print()
box_prefs$n_questions()

head(approx_voters_popn)
popn_obj <- SurveyData$new(
  data = approx_voters_popn,
  questions = list(
    age_group = "Which age group are you?",
    gender = "Gender?",
    vote_pref = "Which party do you prefer to vote for?"
  ),
  # order doesn't matter (gender before age2 here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_voters_popn$gender),
    age_group = levels(approx_voters_popn$age_group),
    vote_pref = levels(approx_voters_popn$vote_pref)
  ),
  weights = "wt"
)
popn_obj$print()

q_age <- QuestionMap$new(
  name = "age",
  col_names = c("age","age_group"),
  values_map = list(
    "18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
    "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"
  )
)
print(q_age)

q_party_pref <- QuestionMap$new(
  name = "party_pref",
  col_names = c("vote_for","vote_pref"),
  values_map = list("Box Party" = "BP",  "BP" = "BP","Circle Party" = "CP", "CP" = "CP")
)
q_gender <- QuestionMap$new(
  name = "gender",
  col_names = c("gender", "gender"),
  values_map = list("male" = "m","female" = "f", "nonbinary" = "nb")
)

# Create SurveyMap object adding all questions at once
ex_map <- SurveyMap$new(
  sample = box_prefs,
  population = popn_obj,
  q_age,
  q_party_pref,
  q_gender
)
print(ex_map) # or ex_map$print()

# Or can add questions incrementally
ex_map <- SurveyMap$new(sample = box_prefs, population = popn_obj)
print(ex_map)

ex_map$add(q_age, q_party_pref)
print(ex_map)

ex_map$add(q_gender)
print(ex_map)

# Create the mapping between sample and population
ex_map$mapping()

# Create the poststratification data frame using all variables in the mapping
# (alternatively, can specify particular variables, e.g. tabulate("age"))
ex_map$tabulate()

# Example rstanarm usage
# Returns a SurveyFit object
fit_1 <- ex_map$fit(
  fun = rstanarm::stan_glmer,
  formula = y ~ (1|age) + (1|gender),
  family = "binomial",
  seed = 1111,
  # just to keep the example fast and small
  chains = 1
)

# Example lme4 usage
# fit_2 <- ex_map$fit(
#   fun = lme4::glmer,
#   formula = y ~ (1|age) + (1|gender),
#   family = "binomial"
# )
#
# Example brms usage
# fit_3 <- ex_map$fit(
#   fun = brms::brm,
#   formula = y ~ (1|age) + (1|gender),
#   family = "bernoulli",
#   seed = 1111
# )


# predicted probabilities
# returns matrix with rows for poststrat cells, cols for posterior draws
poststrat_estimates <- fit_1$population_predict()

# estimates by age level
estimates_by_age <- fit_1$aggregate(poststrat_estimates, by = "age")
head(estimates_by_age)
estimates_by_age \%>\%
  group_by(age) \%>\%
  summarize(mean = mean(value), sd = sd(value))

# plot estimates by age
fit_1$plot(estimates_by_age, weights = FALSE)
fit_1$plot(estimates_by_age, weights = TRUE)

# population estimate
estimates_popn <- fit_1$aggregate(poststrat_estimates)
mean(estimates_popn$value)

# plot population estimate
fit_1$plot(estimates_popn, weights = FALSE)
fit_1$plot(estimates_popn, weights = TRUE)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyMap$new()}}
\item \href{#method-print}{\code{SurveyMap$print()}}
\item \href{#method-add}{\code{SurveyMap$add()}}
\item \href{#method-delete}{\code{SurveyMap$delete()}}
\item \href{#method-replace}{\code{SurveyMap$replace()}}
\item \href{#method-validate}{\code{SurveyMap$validate()}}
\item \href{#method-mapping}{\code{SurveyMap$mapping()}}
\item \href{#method-tabulate}{\code{SurveyMap$tabulate()}}
\item \href{#method-fit}{\code{SurveyMap$fit()}}
\item \href{#method-item_map}{\code{SurveyMap$item_map()}}
\item \href{#method-sample}{\code{SurveyMap$sample()}}
\item \href{#method-population}{\code{SurveyMap$population()}}
\item \href{#method-poststrat_data}{\code{SurveyMap$poststrat_data()}}
\item \href{#method-mapped_sample_data}{\code{SurveyMap$mapped_sample_data()}}
\item \href{#method-mapped_population_data}{\code{SurveyMap$mapped_population_data()}}
\item \href{#method-clone}{\code{SurveyMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SurveyMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$new(sample, population, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{The \link{SurveyData} object corresponding to the sample data.}

\item{\code{population}}{The \link{SurveyData} object corresponding to the population data.}

\item{\code{...}}{\link{QuestionMap} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the mapping.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add new \link{QuestionMap}s.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The \link{QuestionMap}s to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete \link{QuestionMap}s.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$delete(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The \link{QuestionMap}s to delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace"></a>}}
\if{latex}{\out{\hypertarget{method-replace}{}}}
\subsection{Method \code{replace()}}{
Replace one \link{QuestionMap} with another.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$replace(old_question, new_question)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_question}}{The \link{QuestionMap} object to replace.}

\item{\code{new_question}}{The \link{QuestionMap} object to use instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the mapping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapping"></a>}}
\if{latex}{\out{\hypertarget{method-mapping}{}}}
\subsection{Method \code{mapping()}}{
The mapping method uses the given maps between questions to create new sample and population data
that has unified variable names (i.e., if the underlying construct is called \code{age}, both sample and population will have
an \code{age} column, even if in the the raw data both had different variable names). The method also unifies the levels of each
variable in the sample and population so that the maximum set of consistent levels is created. Names of these new levels are
made according the the sample level names. If multiple levels are combined, the new name will be existing levels seperated by a
\code{+}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate}{}}}
\subsection{Method \code{tabulate()}}{
Prepare the poststratification table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$tabulate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The variables to include.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$fit(fun, formula, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{The model fitting function to use. For example,
\code{fun=rstanarm::stan_glmer}, \code{fun=brms::brm}, \code{fun=lme4::glmer}.
If using a custom \code{fun} it must have a \code{data} argument that accepts
a data frame (like standard R modeling functions).}

\item{\code{formula}}{The model formula. Can be either a string or a formula
object.}

\item{\code{...}}{Arguments other than \code{formula} and \code{data} to pass to \code{fun}.
The data argument will be automatically specified internally.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{SurveyFit} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-item_map"></a>}}
\if{latex}{\out{\hypertarget{method-item_map}{}}}
\subsection{Method \code{item_map()}}{
Access the \code{item_map}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$item_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Access the \link{SurveyData} object containing the sample data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-population"></a>}}
\if{latex}{\out{\hypertarget{method-population}{}}}
\subsection{Method \code{population()}}{
Access the \link{SurveyData} object containing the population data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$population()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poststrat_data"></a>}}
\if{latex}{\out{\hypertarget{method-poststrat_data}{}}}
\subsection{Method \code{poststrat_data()}}{
Access the poststratification data frame created by the \code{tabulate} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$poststrat_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_sample_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_sample_data}{}}}
\subsection{Method \code{mapped_sample_data()}}{
Access the data frame containing the mapped sample data
created by the \code{mapping} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapped_sample_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_population_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_population_data}{}}}
\subsection{Method \code{mapped_population_data()}}{
Access the data frame containing the mapped population data
created by the \code{mapping} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapped_population_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
