% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap.R
\name{SurveyMap}
\alias{SurveyMap}
\title{SurveyMap}
\description{
An \link[R6:R6Class]{R6} \code{SurveyMap} object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.  The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\examples{
library(dplyr)

head(feline_survey)
feline_prefs <- SurveyData$new(
  data = feline_survey,
  questions = list(
    age1 = "Please identify your age group",
    gender = "Please select your gender",
    pet_own = "Which pet do you own?",
    y = "Response"
  ),
  responses = list(
    age1 = levels(feline_survey$age1),
    gender = levels(feline_survey$gender),
    pet_own = levels(feline_survey$pet_own),
    y = c("no","yes")
  ),
  weights = feline_survey$wt
)
feline_prefs$print()

head(approx_popn)
popn_obj <- SurveyData$new(
  data = approx_popn,
  questions = list(
    age2 = "Which age group are you?",
    gender = "Gender?",
    pet_pref = "Which pet would you like to own?"
  ),
  # order doesn't matter (gender before age2 here) because
  # the list has the names of the variables
  responses = list(
    gender = levels(approx_popn$gender),
    age2 = levels(approx_popn$age2),
    pet_pref = levels(approx_popn$pet_pref)
  ),
  weights = approx_popn$wt
)
popn_obj$print()

q1 <- SurveyQuestion$new(
  name = "age",
  col_names = c("age1","age2"),
  values_map = list(
    "18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
    "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"
  )
)
q2 <- SurveyQuestion$new(
  name = "pet",
  col_names = c("pet_own","pet_pref"),
  values_map = list("cat" = "cat", "kitten" = "cat","dog" = "dog","puppy" = "dog")
)
q3 <- SurveyQuestion$new(
  name = "gender",
  col_names = c("gender","gender"),
  values_map = data.frame("male" = "m","female" = "f", "nonbinary" = "nb")
)

# Create SurveyMap object
# can add all questions at once or incrementally
ex_map <- SurveyMap$new(sample = feline_prefs, population = popn_obj, q1)
print(ex_map)
ex_map$validate()
ex_map$add(q3)
print(ex_map)
ex_map$delete(q3)
print(ex_map)
ex_map$add(q3)
ex_map$delete("gender")
print(ex_map)
ex_map$add(q2)
print(ex_map)
ex_map$replace(q1,q3)
print(ex_map)
ex_map$add(q1)
print(ex_map)
ex_map$validate()
ex_map$mapping()
ex_map$tabulate("age") # Just use age in the poststrat matrix
ex_map$tabulate() # Use all variables in the map


#' Example rstanarm usage
#' Returns a SurveyFit object
fit_1 <- ex_map$fit(
  fun = rstanarm::stan_glmer,
  formula = y ~ (1|age) + (1|gender),
  family = "binomial",
  refresh = 100,
  cores = 2
)

\dontrun{
# Example brms usage
# Returns a SurveyFit object
# (not run because requires compilation)
fit_2 <- ex_map$fit(
  fun = brms::brm,
  formula = y ~ (1|age) + (1|gender),
  family = "bernoulli",
  refresh = 100,
  cores = 2
)
}

# predicted probabilities
# returns matrix with rows for poststrat cells, cols for posterior draws
poststrat_estimates <- fit_1$population_predict()

# estimates by age level
estimates_by_age <- fit_1$aggregate(poststrat_estimates, by = "age")
head(estimates_by_age)
estimates_by_age \%>\%
  group_by(age) \%>\%
  summarize(mean = mean(value), sd = sd(value))

# plot estimates by age
fit_1$plot(estimates_by_age)

# population estimate
estimates_popn <- fit_1$aggregate(poststrat_estimates)
mean(estimates_popn$value)

# plot population estimate
fit_1$plot(estimates_popn)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyMap$new()}}
\item \href{#method-print}{\code{SurveyMap$print()}}
\item \href{#method-add}{\code{SurveyMap$add()}}
\item \href{#method-delete}{\code{SurveyMap$delete()}}
\item \href{#method-replace}{\code{SurveyMap$replace()}}
\item \href{#method-validate}{\code{SurveyMap$validate()}}
\item \href{#method-mapping}{\code{SurveyMap$mapping()}}
\item \href{#method-tabulate}{\code{SurveyMap$tabulate()}}
\item \href{#method-fit}{\code{SurveyMap$fit()}}
\item \href{#method-item_map}{\code{SurveyMap$item_map()}}
\item \href{#method-sample}{\code{SurveyMap$sample()}}
\item \href{#method-population}{\code{SurveyMap$population()}}
\item \href{#method-poststrat_data}{\code{SurveyMap$poststrat_data()}}
\item \href{#method-mapped_sample_data}{\code{SurveyMap$mapped_sample_data()}}
\item \href{#method-mapped_population_data}{\code{SurveyMap$mapped_population_data()}}
\item \href{#method-clone}{\code{SurveyMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SurveyMap object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$new(sample, population, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{The \link{SurveyData} object corresponding to the sample data.}

\item{\code{population}}{The \link{SurveyData} object corresponding to the population data.}

\item{\code{...}}{\link{SurveyQuestion} objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the mapping.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add new \link{SurveyQuestion}s.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The \link{SurveyQuestion}s to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete \link{SurveyQuestion}s.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$delete(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The \link{SurveyQuestion}s to delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace"></a>}}
\if{latex}{\out{\hypertarget{method-replace}{}}}
\subsection{Method \code{replace()}}{
Replace one \link{SurveyQuestion} with another.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$replace(old_question, new_question)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_question}}{The \link{SurveyQuestion} object to replace.}

\item{\code{new_question}}{The \link{SurveyQuestion} object to use instead.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the mapping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapping"></a>}}
\if{latex}{\out{\hypertarget{method-mapping}{}}}
\subsection{Method \code{mapping()}}{
Prepare the mapped data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate}{}}}
\subsection{Method \code{tabulate()}}{
Prepare the poststratification table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$tabulate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The variables to include.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$fit(fun, formula, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{The model fitting function to use. For example,
\code{fun=rstanarm::stan_glmer}, \code{fun=brms::brm}, \code{fun=lme4::glmer}.
If using a custom \code{fun} it must have a \code{data} argument that accepts
a data frame (like standard R modeling functions).}

\item{\code{formula}}{The model formula. Can be either a string or a formula
object.}

\item{\code{...}}{Arguments other than \code{formula} and \code{data} to pass to \code{fun}.
The data argument will be automatically specified internally.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{SurveyFit} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-item_map"></a>}}
\if{latex}{\out{\hypertarget{method-item_map}{}}}
\subsection{Method \code{item_map()}}{
Access the \code{item_map}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$item_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Access the \link{SurveyData} object containing the sample data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-population"></a>}}
\if{latex}{\out{\hypertarget{method-population}{}}}
\subsection{Method \code{population()}}{
Access the \link{SurveyData} object containing the population data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$population()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poststrat_data"></a>}}
\if{latex}{\out{\hypertarget{method-poststrat_data}{}}}
\subsection{Method \code{poststrat_data()}}{
Access the poststratification data frame created by the \code{tabulate} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$poststrat_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_sample_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_sample_data}{}}}
\subsection{Method \code{mapped_sample_data()}}{
Access the data frame containing the mapped sample data
created by the \code{mapping} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapped_sample_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mapped_population_data"></a>}}
\if{latex}{\out{\hypertarget{method-mapped_population_data}{}}}
\subsection{Method \code{mapped_population_data()}}{
Access the data frame containing the mapped population data
created by the \code{mapping} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$mapped_population_data(key = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Should the \code{.key} column be included? This column just
indicates the original order of the rows and is primarily intended
for internal use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
