% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyMap_alt.R
\docType{data}
\name{SurveyMap_alt}
\alias{SurveyMap_alt}
\alias{question}
\title{SurveyMap}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
question
}
\description{
A `SurveyMap_alt` object holds the mapping
between a set of items in a survey and a population dataset.
The label is the item label in each dataset and the values
is a list of all possible values.   The values for the survey
and population must be aligned, i.e., the lists must have the
same number of elements and the values at index i in each list
are equivalent.  If there is a meaningful ordering over the values,
they should be listed in that order, either descending or ascending.
}
\examples{
survey_values <- c('18-25','26-35','36-45','46-55','56-65','66-75','76-90')
popn_values <- c('18-35','18-35','36-55','36-55','56-65','66+','66+')
q1 <- question$new(name = "age", col_names = c(samp = "age1",popn = "age2"), values = data.frame(age1 = survey_values,age2 = popn_values))
survey_values2 <- c('cat','cat','dog','dog')
popn_values2 <- c('cat','kitten','dog','puppy')
q2 <- question$new(name = "pet", col_names = c(samp = "pet_pref",popn = "pet_own"), values = data.frame(pet_pref = survey_values2,pet_own = popn_values2))
survey_values3 <- c('male','female','nonbinary')
popn_values3 <- c('m','f','nb')
q3 <- question$new(name = "gender", col_names = c(samp = "gender",popn = "gender"), values = data.frame(gender = survey_values3,gender = popn_values3))

tmp_map <- SurveyMap_alt$new(q1,q2)
print(tmp_map)
tmp_map$add(q3)
print(tmp_map)
tmp_map$delete(q3)
print(tmp_map)
tmp_map$replace(q1,q3)
print(tmp_map)
}
\keyword{datasets}
