% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyObj.R
\docType{data}
\name{SurveyObject}
\alias{SurveyObject}
\alias{SurveyObj}
\title{SurveyObject}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SurveyObj
}
\description{
A `SurveyObject` represents a survey and its metadata.
The survey itself is a dataframe.
The survey metatdata consists of
 - per-column questions - a list of strings
 - per-column allowed response values - a list of lists
 - per-column survey weights, - a vector of numeric weights
 - a string that specifies the survey design using lmer forumula syntax

Examples of survey design include:
~. a random sample, ~ (1|cluster), a one stage cluster sample, ~ stratum, a stratified sample
}
\examples{
feline_prefs = SurveyObj$new(feline_survey[,c("age1","gender","pet_own","y")],
questions = c("Please identify your age group","Please select your gender","Which pet do you own?", "Response"),
responses = list(levels(feline_survey$age1),levels(feline_survey$gender),levels(feline_survey$pet_own),c("no","yes")),
weights = feline_survey$wt,
design = formula("~."))

popn_obj = SurveyObj$new(approx_popn[,c("age2","gender","pet_pref")],
questions = c("Which age group are you?","Gender?","Which pet would you like to own?"),
responses = list(levels(approx_popn$age2),levels(approx_popn$gender),levels(approx_popn$pet_pref)),
weights = approx_popn$wt,
design = formula("~."))
}
\keyword{datasets}
