% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyQuestion.R
\name{SurveyQuestion}
\alias{SurveyQuestion}
\title{SurveyQuestion}
\description{
An \link[R6:R6Class]{R6} \code{SurveyQuestion} object holds the mapping
for one question or demographic between the survey and population dataset.
}
\examples{
q1 <- SurveyQuestion$new(
  name = "age",
  col_names = c("age1","age2"),
  values_map = list(
    "18-25" = "18-35", "26-35" = "18-35","36-45" = "36-55",
    "46-55" = "36-55", "56-65" = "56-65", "66-75" = "66+", "76-90" = "66+"
  )
)
print(q1)

q2 <- SurveyQuestion$new(
  name = "pet",
  col_names = c("pet_pref","pet_own"),
  values_map = list("cat" = "cat", "kitten" = "cat","dog" = "dog","dog" = "puppy")
)
q3 <- SurveyQuestion$new(
  name = "gender",
  col_names = c("gender", "gender"),
  values_map = list("male" = "m","female" = "f", "nonbinary" = "nb")
)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SurveyQuestion$new()}}
\item \href{#method-print}{\code{SurveyQuestion$print()}}
\item \href{#method-name}{\code{SurveyQuestion$name()}}
\item \href{#method-col_names}{\code{SurveyQuestion$col_names()}}
\item \href{#method-values}{\code{SurveyQuestion$values()}}
\item \href{#method-clone}{\code{SurveyQuestion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SurveyQuestion}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyQuestion$new(name, col_names, values_map)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the underlying construct. For example if the
sample data uses the name \code{age1} and the population data uses
the name \code{age2} then \code{name} could be \code{age}.}

\item{\code{col_names}}{A character vector of length two giving the column names
of the question in the sample data (first element) and the population
data (second element).}

\item{\code{values_map}}{A named list where the names correspond to the
responses to the question in the sample data and the values correspond
to the responses in the population data. If there is a meaningful
ordering over the values, they should be listed in that order, either
descending or ascending. See \strong{Examples}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{SurveyQuestion} object that can be added to a \code{\link{SurveyMap}}
object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{SurveyQuestion}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyQuestion$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \code{SurveyQuestion} object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\if{latex}{\out{\hypertarget{method-name}{}}}
\subsection{Method \code{name()}}{
Access \code{name}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyQuestion$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-col_names"></a>}}
\if{latex}{\out{\hypertarget{method-col_names}{}}}
\subsection{Method \code{col_names()}}{
Access \code{col_names}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyQuestion$col_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-values"></a>}}
\if{latex}{\out{\hypertarget{method-values}{}}}
\subsection{Method \code{values()}}{
Access \code{values}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyQuestion$values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveyQuestion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
